Move opaque type cache into `InferCtxt`
